Facade Design Pattern
The façade pattern is commonly implemented for the following purposes
and situations:
- Provide a simple and uni ed access to a legacy back‐end system.
Create a public API to classes, such as a driver.
- Offer coarse‐grained access to available services. Services are
combined, such as in the washing machine example.
- Reduce network calls. The façade makes many calls to the subsystem,
while the remote client makes only one call to the façade.
- To encapsulate  ow and inner details of an app for security and
simplicity.

 NOTE: Façades are sometimes implemented as singleton abstract
factories.

The benefits of the façade pattern:

- A reduction in coupling because the client knows nothing about the
subsystem
Increased maintainability and manageability when changes are required
- Reuse of functionality because it encourages the reuse of controls
and  fine‐grained logic
- Consistency of service execution by invoking the same method
consistently from one invocation to the next
- Reduction in business logic complexity by grouping related methods
and invoking them from one method invocation
- Centralization of security and transaction control management
- Testable and mockable pattern implementations

WHERE AND WHEN TO USE THE FAÇADE PATTERN

The façade pattern should be used to encapsulate complicated
(business) logic at a high level and provide a cleaner single point of
access via an API.

Whenever you are in the position to provide an interface or an API to
someone, think  rst about the complexity of the logic and the changes
that might occur. The façade pattern does a good job of providing a
clean API while hiding the parts subject to change.
However, unnecessarily wrapping methods in a façade is bad practice
and adds unnecessary layers. Premature encapsulation could result in
too many invocations and layers that don’t add value.
When implementing the session façade, you must determine if the use
case requires state to be maintained. A use case that invokes only one
method of the façade to receive the service that it needs is
considered non conversational, so there is no need to save the
conversational state between one method invocation and the next. You
should implement this façade as a stateless session bean.

On the other hand, if the conversational state must be maintained
between method invocations, the most appropriate way to implement the
façade is as a stateful session bean.

You must be careful in the use of the stateful session façade because
it ties up server resources until the client that provoked the
conversation releases them or times out. This could mean that, for the
majority of the time the stateful session bean façade is bound to the
client, it is doing nothing; it’s just maintaining state and using
resources. And, unlike stateless session bean façades, it cannot be
reused and shared between other clients because each request creates a
new instance of the stateless façade, maintaining the state for that
client’s session.

So take care when using this pattern. Analyze the use case and make
appropriate decisions.